Add detailsPage
